OCPBUGS-81319: Remove dynamic-core-api
We had two files and now we have one
